requirements.txt
ephem
exif
h5py==2.10.0
keras==2.2.5
logzero
matplotlib
numpy
opencv-python
pillow
reverse_geocoder
tensorflow==1.14

